feat: Disable new versions if there is already a draft
* Add is newest_flag to report chain to confirm whether newer versions exist, but dont return their data
* Use this flag to control the supplemental and assessment buttons with dialogs + text
