+++ backend/lcfs/tests/fuel_export/test_fuel_exports_views.py
@@ -8,9 +8,6 @@ from unittest.mock import patch
 
 from lcfs.db.models import ComplianceReport
 from lcfs.db.models.user.Role import RoleEnum
-from lcfs.tests.compliance_report.conftest import (
-    compliance_report_base_schema,
-)
 from lcfs.web.api.fuel_export.schema import (
     FuelExportSchema,
     FuelExportCreateUpdateSchema,
